codecs: rename files for consistency
